+++ Sources/LogView/NavigationController.swift
@@ -1,0 +1,16 @@
+//
+//  NavigationController.swift
+//  Guppy
+//
+//  Created by Stephen Hayes on 12/17/19.
+//  Copyright © 2019 Johnson & Johnson. All rights reserved.
+//
+
+import UIKit
+
+class NavigationController: UINavigationController {
+    
+    override var preferredStatusBarStyle: UIStatusBarStyle {
+        return .lightContent
+    }
+}